Enable apply_filter to handle an array or ActiveRecord::Relation
The change from using `records_for` to `public_send(@relation_type)` is causing errors that are currently untested because `apply_filter` expected `records` to be an `ActiveRecord::Relation` but it's now sometimes an `Array`.

This commit does not include a test, but I'll figure out how to add one once I confirm that it works in an application that found the error.